+++ apps/web/src/app/[locale]/(admin)/poll/[urlId]/overview-page.tsx
@@ -2,21 +2,20 @@
 
 import { cn } from "@rallly/ui";
 import { Button } from "@rallly/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@rallly/ui/card";
+import { Card, CardContent } from "@rallly/ui/card";
 import { Icon } from "@rallly/ui/icon";
+import { Label } from "@rallly/ui/label";
 import { ArrowUpRightIcon, LinkIcon } from "lucide-react";
 import Link from "next/link";
 import React from "react";
 import { useCopyToClipboard } from "react-use";
 
+import Discussion from "@/components/discussion";
 import { EventCard } from "@/components/event-card";
 import { useTouchBeacon } from "@/components/poll/use-touch-beacon";
+import { VotingForm } from "@/components/poll/voting-form";
+import { PollViz } from "@/components/poll-viz";
+import { ProBadge } from "@/components/pro-badge";
 import { Trans } from "@/components/trans";
 import { usePoll } from "@/contexts/poll";
 
@@ -61,7 +60,6 @@ export function ShareCard() {
   const poll = usePoll();
   return (
     <div className="space-y-4">
-      <EventCard />
       <div>
         <div className="flex gap-x-2.5">
           <div className="grow">
@@ -85,7 +83,19 @@ export function OverviewPage() {
   useTouchBeacon(poll.id);
   return (
     <div className="space-y-4 lg:space-y-6">
-      <ShareCard />
+      <div className="grid grid-cols-3 gap-4">
+        <Button size="lg">Share</Button>
+        <Button size="lg">
+          Invite Participants <ProBadge />
+        </Button>
+        <Button size="lg" asChild>
+          <Link href={`/poll/${poll.id}/finalize`}>
+            Outcome <ProBadge />
+          </Link>
+        </Button>
+      </div>
+
+      <Discussion />
     </div>
   );
 }
